refactor: change interfaces location
