+++ src/app/components/shared/MyRatingNumber/MyRatingNumber.tsx
@@ -5,7 +5,7 @@ import { IconStar } from '@/app/components/shared/icons/IconStar';
 import { Group, Text, UnstyledButton } from '@mantine/core';
 import classes from '@/app/components/shared/MyRatingNumber/MyRatingNumber.module.css';
 import DataLocalStorageProvider from '@/lib/services/DataLocalStorageProvider';
-import { IUserMovieRating } from '@/interfaces/IUserMovieRating';
+import { IUserMovieRating } from '@/types/interfaces/IUserMovieRating';
 import { UserRatingModal } from '@/app/components/shared/UserRatingModal/UserRatingModal';
 import { useEffect, useState } from 'react';
