+++ src/types/interfaces/ILocalConfigs.ts
