+++ src/types/interfaces/IUserData.ts
@@ -1,4 +1,4 @@
-import { IUserMovieRating } from './IUserMovieRating';
+import { IUserMovieRating } from '@/types/interfaces/IUserMovieRating';
 
 export interface IUserData {
   userMoviesRating: Array<IUserMovieRating>;
